Refactored the input handling and Create function code
