+++ samples/MauiMsApp/CreatePage.xaml.cs
@@ -13,38 +13,6 @@ public partial class CreatePage : ContentPage
         InitializeComponent();
     }
 
-
-    /// <summary>
-    /// Attempt to do specified app creation
-    /// </summary>
-    private async Task CreateMSApp(bool interactive, string fullPathToMsApp, int numScreens, IList<string>? controlsinfo)
-    {
-        try
-        {
-            // Setup services for creating MSApp representation
-            var provider = Handler!.MauiContext!.Services;
-
-            // Create a new empty MSApp
-            using var msapp = provider.GetRequiredService<IMsappArchiveFactory>().Create(fullPathToMsApp);
-
-            // Select Generator based off specified mode
-            var generator = provider.GetRequiredService<IAppGeneratorFactory>().Create(interactive);
-
-            // Generate the app
-            msapp.App = generator.GenerateApp(Path.GetFileNameWithoutExtension(fullPathToMsApp),
-                    numScreens, controlsinfo);
-
-            // Output the MSApp to the path provided
-            msapp.Save();
-
-            await DisplayAlert("Success", "You are now a PowerApps Pro Developer!", "OK");
-        }
-        catch (Exception ex)
-        {
-            await DisplayAlert("Alert", "Something went wrong:\n" + ex.Message, "OK");
-        }
-    }
-
 #pragma warning disable CA1822 // Mark members as static
     private async void OnCreateClicked(object sender, EventArgs e)
 #pragma warning restore CA1822 // Mark members as static
@@ -55,11 +23,16 @@ public partial class CreatePage : ContentPage
             int numScreens;
             var result = int.TryParse(_numScreensEntry.Text, out numScreens);
             var controlTemplates = _controlTemplatesEntry.Text.Split(' ');
-            await CreateMSApp(false, filePath, numScreens, controlTemplates);
+
+            var creator = new AppCreator(Handler!.MauiContext!.Services);
+            creator.CreateMSApp(false, filePath, numScreens, controlTemplates);
+
+            await DisplayAlert("Success", "You are now a PowerApps Pro Developer!", "OK");
         }
-        catch (Exception)
+        catch (Exception ex)
         {
             // The user canceled or something went wrong
+            await DisplayAlert("Alert", "Something went wrong: " + ex.Message, "OK");
         }
     }
